third_party/blink/web_tests/platform/generic/fast/forms/radionodelist-whose-form-element-detached-from-domtree-expected.txt

RadioNodeList should be rooted at form itself if detached from dom tree

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



form in dom tree
PASS radioNodeList1.length is 3
PASS radioNodeList2.length is 3
PASS radioNodeList1[0].value is 'value1'
PASS radioNodeList1[1].value is 'value2'
PASS radioNodeList1[2].value is 'value3'
PASS radioNodeList2[0].value is 'value4'
PASS radioNodeList2[1].value is 'value5'
PASS radioNodeList2[2].value is 'value6'
Check RadioNodeList.value
PASS radioNodeList1.value is ""
PASS radioNodeList2.value is ""
PASS radioNodeList1[2].checked = true; radioNodeList1.value is 'value3'


form detached from dom tree
PASS radioNodeList1.length is 2
PASS radioNodeList2.length is 2
PASS radioNodeList1[0].value is 'value1'
PASS radioNodeList1[1].value is 'value2'
PASS radioNodeList2[0].value is 'value4'
PASS radioNodeList2[1].value is 'value5'
Check RadioNodeList.value
PASS radioNodeList1[1].checked = true; radioNodeList1.value is 'value2'


form again added to dom tree
PASS radioNodeList1.length is 3
PASS radioNodeList2.length is 3
PASS radioNodeList1[0].value is 'value1'
PASS radioNodeList1[1].value is 'value2'
PASS radioNodeList1[2].value is 'value3'
PASS radioNodeList2[0].value is 'value4'
PASS radioNodeList2[1].value is 'value5'
PASS radioNodeList2[2].value is 'value6'


form's ancestor detached from dom tree
PASS radioNodeList1.length is 2
PASS radioNodeList2.length is 2
PASS radioNodeList1[0].value is 'value1'
PASS radioNodeList1[1].value is 'value2'
PASS radioNodeList2[0].value is 'value4'
PASS radioNodeList2[1].value is 'value5'
Check RadioNodeList.value
PASS radioNodeList1[1].checked = true; radioNodeList1.value is 'value2'


form again added to dom tree
PASS radioNodeList1.length is 3
PASS radioNodeList2.length is 3
PASS radioNodeList1[0].value is 'value1'
PASS radioNodeList1[1].value is 'value2'
PASS radioNodeList1[2].value is 'value3'
PASS radioNodeList2[0].value is 'value4'
PASS radioNodeList2[1].value is 'value5'
PASS radioNodeList2[2].value is 'value6'
PASS successfullyParsed is true

TEST COMPLETE
